server.rb only proccesses applications
