+++ sinatra-backend/Gemfile
@@ -3,3 +3,4 @@ source 'https://rubygems.org'
 gem 'sinatra'
 gem 'sinatra-contrib'
 gem 'google_drive', '~> 1.0.1'
+gem 'pry'
