+++ sinatra-backend/views/payment.erb
@@ -47,14 +47,14 @@
       <div class="row">
         <div class="small-10 small-centered large-4 large-centered columns">
           <div class="panel">
-            <p>Venue details here...</p>
+            <p></p>
           </div>
         </div>
       </div>
     </div>
 
     <div class="section red text-center">
-      <h1 class="white open-sans">Application payment</h1>
+      <h1 class="open-sans">Application payment</h1>
       <div class="row">
         <div class="small-10 small-centered large-4 large-centered columns">
             <form action="" method="POST">
